scripts: rename scr_environment file and function names
