+++ scripts/python/scrjob/scr_list_dir.py
@@ -6,7 +6,7 @@ import argparse
 
 from scrjob import scr_const
 from scrjob.list_dir import list_dir
-from scrjob.scr_environment import SCR_Env
+from scrjob.environment import SCR_Env
 from scrjob.scr_param import SCR_Param
 from scrjob.resmgrs import AutoResourceManager
