+++ scripts/python/scrjob/scr_scavenge.py
@@ -9,7 +9,7 @@ from time import time
 
 from scrjob import scr_const
 from scrjob.scr_param import SCR_Param
-from scrjob.scr_environment import SCR_Env
+from scrjob.environment import SCR_Env
 from scrjob.resmgrs import AutoResourceManager
 from scrjob.launchers import AutoJobLauncher
